# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "yes",
        "title": "twitchclone api",
        "termsOfService": "no",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8090",
    "basePath": "/api",
    "paths": {
        "/categories": {
            "get": {
                "description": "Retrieves a list of categories with optional pagination and sorting parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "List categories with pagination and sorting",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of items per page",
                        "name": "count",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "description": "Sort order (asc or desc)",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of categories",
                        "schema": {
                            "$ref": "#/definitions/category.ListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid query parameter",
                        "schema": {
                            "$ref": "#/definitions/er.RequestError"
                        }
                    },
                    "500": {
                        "description": "Internal server error while fetching categories",
                        "schema": {
                            "$ref": "#/definitions/er.RequestError"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new category with the given details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "Add a new category",
                "parameters": [
                    {
                        "description": "Category creation request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/category.PostRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content - category created successfully"
                    },
                    "400": {
                        "description": "Bad Request: invalid data",
                        "schema": {
                            "$ref": "#/definitions/er.RequestError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: failed to add category",
                        "schema": {
                            "$ref": "#/definitions/er.RequestError"
                        }
                    }
                }
            }
        },
        "/categories/{categoryIdentifier}": {
            "get": {
                "description": "Retrieves detailed information about a category by either its numeric ID or a unique link (slug).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "Get category by ID or link",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category identifier (either numeric ID or unique link)",
                        "name": "categoryIdentifier",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category found successfully",
                        "schema": {
                            "$ref": "#/definitions/category.GetResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/er.RequestError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a category entity identified by either its numeric ID or unique link (slug).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "Delete category by ID or link",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category identifier (either numeric ID or unique link)",
                        "name": "categoryIdentifier",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content - category delted successfully"
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/er.RequestError"
                        }
                    }
                }
            },
            "patch": {
                "description": "Partially updates a category entity using provided fields, identified by either numeric ID or unique link (slug).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "Update category by ID or link",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category identifier (either numeric ID or unique link)",
                        "name": "categoryIdentifier",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update in the category",
                        "name": "patchRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/category.PatchRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content - category updated successfully"
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/er.RequestError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/er.RequestError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "category.Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "isSafe": {
                    "type": "boolean"
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                },
                "viewers": {
                    "type": "integer"
                }
            }
        },
        "category.GetResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/category.Category"
                }
            }
        },
        "category.ListResponse": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/category.ListResponseItem"
                    }
                }
            }
        },
        "category.ListResponseItem": {
            "type": "object",
            "properties": {
                "is_safe": {
                    "type": "boolean"
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "thumbnail": {
                    "type": "string"
                },
                "viewers": {
                    "type": "integer"
                }
            }
        },
        "category.PatchRequest": {
            "type": "object",
            "properties": {
                "is_safe": {
                    "type": "boolean"
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "thumbnail": {
                    "type": "string"
                }
            }
        },
        "category.PostRequest": {
            "type": "object",
            "properties": {
                "is_safe": {
                    "type": "boolean"
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "thumbnail": {
                    "type": "string"
                }
            }
        },
        "er.RequestError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}